requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
psycopg2-binary
